+++ backend/reviewduck/src/main/java/com/reviewduck/config/datasource/CustomDatasourceConfig.java
@@ -1,0 +1,73 @@
+package com.reviewduck.config.datasource;
+
+import static com.reviewduck.config.datasource.DataSourceConfiguration.*;
+
+import java.util.HashMap;
+
+import javax.persistence.EntityManagerFactory;
+import javax.sql.DataSource;
+
+import org.springframework.boot.autoconfigure.orm.jpa.JpaProperties;
+import org.springframework.boot.orm.jpa.EntityManagerFactoryBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.datasource.LazyConnectionDataSourceProxy;
+import org.springframework.orm.jpa.JpaTransactionManager;
+import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
+import org.springframework.orm.jpa.vendor.AbstractJpaVendorAdapter;
+import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
+import org.springframework.transaction.PlatformTransactionManager;
+
+@Configuration
+public class CustomDatasourceConfig {
+    private final ReplicationRoutingDataSource routingDataSource;
+    private final JpaProperties jpaProperties;
+
+    public CustomDatasourceConfig(ReplicationRoutingDataSource routingDataSource,
+        JpaProperties jpaProperties) {
+        this.routingDataSource = routingDataSource;
+        this.jpaProperties = jpaProperties;
+    }
+
+    @Bean
+    public DataSource routingDataSource(
+    ) {
+        DataSourceConfiguration dataSourceConfiguration = new DataSourceConfiguration();
+        DataSource sourceDataSource = dataSourceConfiguration.sourceDataSource();
+        DataSource replicaDataSource = dataSourceConfiguration.replicaDataSource();
+
+        HashMap<Object, Object> dataSourceMap = new HashMap<>();
+        dataSourceMap.put(SOURCE, sourceDataSource);
+        dataSourceMap.put(REPLICA, replicaDataSource);
+
+        routingDataSource.setTargetDataSources(dataSourceMap);
+        routingDataSource.setDefaultTargetDataSource(sourceDataSource);
+
+        return routingDataSource;
+    }
+
+    @Bean(name = "dataSource")
+    public DataSource dataSource() {
+        return new LazyConnectionDataSourceProxy(routingDataSource());
+    }
+
+    @Bean
+    public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
+        EntityManagerFactoryBuilder entityManagerFactoryBuilder = createEntityManagerFactoryBuilder(jpaProperties);
+        return entityManagerFactoryBuilder.dataSource(dataSource())
+            .packages("com.reviewduck")
+            .build();
+    }
+
+    private EntityManagerFactoryBuilder createEntityManagerFactoryBuilder(JpaProperties jpaProperties) {
+        AbstractJpaVendorAdapter vendorAdapter = new HibernateJpaVendorAdapter();
+        return new EntityManagerFactoryBuilder(vendorAdapter, jpaProperties.getProperties(), null);
+    }
+
+    @Bean
+    public PlatformTransactionManager transactionManager(EntityManagerFactory entityManagerFactory) {
+        JpaTransactionManager tm = new JpaTransactionManager();
+        tm.setEntityManagerFactory(entityManagerFactory);
+        return tm;
+    }
+}
