+++ backend/reviewduck/src/main/java/com/reviewduck/config/datasource/ReplicationRoutingDataSource.java
@@ -2,37 +2,14 @@ package com.reviewduck.config.datasource;
 
 import static com.reviewduck.config.datasource.DataSourceConfiguration.*;
 
-import java.util.HashMap;
-
-import javax.persistence.EntityManagerFactory;
-import javax.sql.DataSource;
-
-import org.hibernate.jpa.boot.spi.EntityManagerFactoryBuilder;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.autoconfigure.orm.jpa.JpaProperties;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.datasource.LazyConnectionDataSourceProxy;
 import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-import org.springframework.orm.jpa.vendor.AbstractJpaVendorAdapter;
-import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
-import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 @Configuration
 public class ReplicationRoutingDataSource extends AbstractRoutingDataSource {
 
     private final DataSourceSelector dataSourceSelector = new DataSourceSelector(REPLICA, SOURCE);
-
-    private final ReplicationRoutingDataSource routingDataSource;
-    private final JpaProperties jpaProperties;
-
-    public ReplicationRoutingDataSource(ReplicationRoutingDataSource routingDataSource, JpaProperties jpaProperties) {
-        this.routingDataSource = routingDataSource;
-        this.jpaProperties = jpaProperties;
-    }
 
     @Override
     protected Object determineCurrentLookupKey() {
@@ -42,46 +19,5 @@ public class ReplicationRoutingDataSource extends AbstractRoutingDataSource {
             return dataSourceSelector.getOne();
         }
         return SOURCE;
-    }
-
-    @Bean
-    public DataSource routingDataSource(
-        @Qualifier(SOURCE) DataSource sourceDataSource,
-        @Qualifier(REPLICA) DataSource replicaDataSource
-    ) {
-
-        HashMap<Object, Object> dataSourceMap = new HashMap<>();
-        dataSourceMap.put(SOURCE, sourceDataSource);
-        dataSourceMap.put(REPLICA, replicaDataSource);
-
-        routingDataSource.setTargetDataSources(dataSourceMap);
-        routingDataSource.setDefaultTargetDataSource(sourceDataSource);
-
-        return routingDataSource;
-    }
-
-    @Bean(name = "dataSource")
-    public DataSource dataSource(@Qualifier("routingDataSource") DataSource routingDataSource) {
-        return new LazyConnectionDataSourceProxy(routingDataSource);
-    }
-
-    @Bean
-    public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
-        EntityManagerFactoryBuilder entityManagerFactoryBuilder = createEntityManagerFactoryBuilder(jpaProperties);
-        return entityManagerFactoryBuilder.dataSource(dataSource())
-            .packages("com.reviewduck")
-            .build();
-    }
-
-    private EntityManagerFactoryBuilder createEntityManagerFactoryBuilder(JpaProperties jpaProperties) {
-        AbstractJpaVendorAdapter vendorAdapter = new HibernateJpaVendorAdapter();
-        return new EntityManagerFactoryBuilder(vendorAdapter, jpaProperties.getProperties(), null);
-    }
-
-    @Bean
-    public PlatformTransactionManager transactionManager(EntityManagerFactory entityManagerFactory) {
-        JpaTransactionManager tm = new JpaTransactionManager();
-        tm.setEntityManagerFactory(entityManagerFactory);
-        return tm;
     }
 }
